+++ CHANGELOG.md
@@ -7,6 +7,13 @@ This file keeps track of all changes to this project. This project follows [sema
 
 ## [UNRELEASED]
 
+### Added
+- Convenient builder for `Result`s: `de.fakeller.performance.analysis.result.Attach`.
+
+### Changed
+- `BusyTime`/`IdleTime` do not require a time unit anymore. Times are in milliseconds by default.
+- Convenience methods in `PerformanceResultWriter` have been removed. Use `Attach` builder to create results instead.
+
 
 ## v0.2.0 - 2016-10-18